requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
pydantic[email]
PyJWT
passlib==1.7.4
bcrypt==3.2.0
python-multipart
pandas
statsmodels